Fixes to unit tests for the System.Convert class.
Simple remark and set the IsSigned to false for the UInt64 convert method.  Just documentation since IsSigned would be set to false in the native code anyhow.

Removed and changed any tests that required the ToString() methods for double and float to handle their max values.  Most of the tests now look at the actual double/float as stored which uses the IEEE 574 format.  Hex display is used for actual asserts.

Added message lines to tests as needed to be able to tell which test was actually failing.

Fixes #715