Clean up unit tests
Added a new FormatException type to match .net core for errors in parsing to numbers.  In a few cases upgraded from an Exception to an ArgumentOutOfRangeException to match .net core.

Removed unneeded debug statements from tests that are now fixed.  In many cases added messages to tests so we can identify the soruce of a failed test.

Upgraded version of CoreLibrary to account for new FormatException type.  Goes with matching nanoCLR with new checksums and new version for mscorlib.

Fix #708